requirements.txt
# Core dependencies for YOLOv5 on Jetson Nano
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
PyYAML>=5.3.1
scipy>=1.5.0
tqdm>=4.41.0
matplotlib>=3.2.2
seaborn>=0.11.0

# YOLOv5 specific
ultralytics

# Audio feedback removed (disabled for now)

# Optional: For TensorRT optimization (install separately on Jetson)
# tensorrt
# pycuda
